Day 22. It's always a pain in the reactor.

+++ day06/src/main.rs
@@ -1,6 +1,6 @@
 fn main() {
-    let mut lanternfish = vec![0 as usize; 9];
-    for fish in std::fs::read_to_string("./input.txt").unwrap().split(",").map(|x| x.parse::<usize>().unwrap()){
+    let mut lanternfish = vec![0_usize; 9];
+    for fish in std::fs::read_to_string("./input.txt").unwrap().split(',').map(|x| x.parse::<usize>().unwrap()){
         lanternfish[fish] += 1;
     }
     for _ in 0..256 {
